refactor: View 관련 기능을 모두 Presentaional으로 이관한다
