+++ src/js/components/LottoDetailContainer/index.js
@@ -1,20 +1,10 @@
-import { JS_SELECTOR } from "../../constants/index.js";
 import store from "../../store/index.js";
-import { $, toDataAttributeSelector as toDAS } from "../../utils/index.js";
 import Presentational from "./Presentational.js";
 
 const createContainer = () => {
-  const $lottoIconWrapper = $(toDAS(JS_SELECTOR.LOTTO_DETAIL.ICON_WRAPPER));
-  const $lottoDetailContainer = $(toDAS(JS_SELECTOR.LOTTO_DETAIL.CONTAINER));
-
   const select = (state) => state.lottos;
 
   let currentLottos = select(store.getState());
-
-  const toggleDetailMode = (force) => {
-    $lottoIconWrapper.toggle("flex-col", force);
-    $lottoDetailContainer.toggle("detail", force);
-  };
 
   const render = () => {
     const previousLottos = currentLottos;
@@ -26,12 +16,11 @@ const createContainer = () => {
     Presentational.render({
       lottos: currentLottos,
       isLottoCleared: currentLottos.length === 0,
-      toggleDetailMode,
     });
   };
 
   const init = () => {
-    Presentational.init({ toggleDetailMode });
+    Presentational.init();
     store.subscribe(render);
   };
