+++ src/js/components/Modal/index.js
@@ -1,13 +1,10 @@
-import { ACTION_TYPE, CLASSNAME } from "../../constants/index.js";
+import { ACTION_TYPE } from "../../constants/index.js";
 import { Lotto } from "../../models/index.js";
 import store from "../../store/index.js";
-import { $, toClassSelector as toCS } from "../../utils/index.js";
 import Presentaional from "./Presentational.js";
 
 const createModal = () => {
   const WINNING_MONEY_UNITS = [2e9, 30e6, 1.5e6, 50e3, 5e3, 0];
-
-  const $container = $(toCS(CLASSNAME.MODAL));
 
   const isRanked = {
     First: ({ matchCount }) => matchCount === 6,
@@ -101,12 +98,8 @@ const createModal = () => {
     store.dispatch({ type: ACTION_TYPE.CLEAR });
   };
 
-  const closeModal = () => {
-    $container.classList.remove(CLASSNAME.MODAL.OPEN);
-  };
-
   const init = () => {
-    Presentaional.init({ closeModal, restart });
+    Presentaional.init({ restart });
 
     store.subscribe(render);
   };
